Move CompareUtils to base src folder and make link
